+++ oracle/price-feeder/oracle/provider/gate.go
@@ -126,7 +126,11 @@ func NewGateProvider(
 	}
 
 	wsConn, response, err := websocket.DefaultDialer.Dial(wsURL.String(), nil)
-	defer response.Body.Close()
+	defer func() {
+		if response != nil {
+			response.Body.Close()
+		}
+	}()
 	if err != nil {
 		return nil, fmt.Errorf("error connecting to Gate websocket: %w", err)
 	}
@@ -551,7 +555,11 @@ func (p *GateProvider) reconnect() error {
 
 	p.logger.Debug().Msg("reconnecting websocket")
 	wsConn, response, err := websocket.DefaultDialer.Dial(p.wsURL.String(), nil)
-	defer response.Body.Close()
+	defer func() {
+		if response != nil {
+			response.Body.Close()
+		}
+	}()
 	if err != nil {
 		return fmt.Errorf("error reconnecting to Gate websocket: %w", err)
 	}