+++ oracle/price-feeder/oracle/provider/huobi.go
@@ -116,7 +116,11 @@ func NewHuobiProvider(
 	}
 
 	wsConn, response, err := websocket.DefaultDialer.Dial(wsURL.String(), nil)
-	defer response.Body.Close()
+	defer func() {
+		if response != nil {
+			response.Body.Close()
+		}
+	}()
 	if err != nil {
 		return nil, fmt.Errorf("error connecting to Huobi websocket: %w", err)
 	}
@@ -388,7 +392,11 @@ func (p *HuobiProvider) reconnect() error {
 
 	p.logger.Debug().Msg("reconnecting websocket")
 	wsConn, response, err := websocket.DefaultDialer.Dial(p.wsURL.String(), nil)
-	defer response.Body.Close()
+	defer func() {
+		if response != nil {
+			response.Body.Close()
+		}
+	}()
 	if err != nil {
 		return fmt.Errorf("error reconnecting to Huobi websocket: %w", err)
 	}